requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.7
requests==2.31.0
python-dotenv==1.0.0
pillow==9.4.0
faker==18.9.0
aiohttp==3.8.5
aiofiles==23.1.0
python-dateutil==2.8.2
numpy==1.24.4
scipy==1.11.2
duckdb==0.8.1
pytesseract==0.3.10
markdown==3.4.4
httpx==0.24.1
beautifulsoup4==4.12.2
pydub==0.25.1
speechrecognition==3.10.0
uv==1.44.1